Properly integrate cargo into the current Makefile system
