--- Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "riscv-pk-rs"
-version = "0.1.0"
-edition = "2018"
-
-[lib]
-path = "lib.rs"
-crate-type = [ "staticlib" ]
-
-[dependencies]
-sm = { path="./sm" }
-util = { path="./util" }
