+++ Makefile.in
@@ -96,7 +96,7 @@ COMPILE       := $(CC) -MMD -MP $(CFLAGS) \
 
 LD            := $(CC)
 LDFLAGS       := @LDFLAGS@ -nostartfiles -nostdlib -static $(LDFLAGS)
-LIBS          := @LIBS@ librs.a
+LIBS          := @LIBS@
 LINK          := $(LD) $(LDFLAGS)
 
 # Library creation
@@ -149,23 +149,6 @@ define reverse_list_h
 endef
 
 
-librs_out := librs/riscv64gc-unknown-none-elf/debug
-librs.a : $(librs_out)/riscv_pk_rs.S
-	$(COMPILE) -c $^ 
-	$(AR) rcv $@ riscv_pk_rs.o
-	$(RANLIB) $@
-
-$(librs_out)/riscv_pk_rs.S: $(librs_out)/riscv_pk_rs.a
-	mkdir librs || true
-	$(eval _PWD := $(shell pwd))
-	cd $(src_dir) ;\
-	export CFLAGS=$(PLATFORM_FLAG) ;\
-	cargo rustc --target-dir $(_PWD)/librs -- --emit=llvm-bc -Clto
-	llc -march=riscv64 -target-abi=lp64d -mattr="+64bit,+m,+a,+f,+d,+c" -code-model=medium $(librs_out)/deps/*.bc -o $(librs_out)/riscv_pk_rs.S
-
-$(librs_out)/riscv_pk_rs.a:
-
--include $(librs_out)/libriscv_pk_rs.d
 
 #-------------------------------------------------------------------------
 # Template for per subproject rules
@@ -209,11 +192,33 @@ $$($(2)_c_objs) : %.o : %.c
 $$($(2)_asm_objs) : %.o : %.S
 	$(COMPILE) -c $$<
 
+
+$(2)_crate_asm := $$(addsuffix .S, $$($(2)_crates))
+$(2)_crate_objs := $$(patsubst %.S, %.o, $$($(2)_crate_asm))
+$(2)_crate_libs := $$(patsubst %.o, %.a, $$($(2)_crate_objs))
+
+.PHONY: $$($(2)_crates)
+
+$$($(2)_crate_asm) : %.S : %
+	mkdir -p $$<
+	$(eval _PWD := $(shell pwd))
+	cd $(src_dir)/$(2) ;\
+	  export CFLAGS=$(PLATFORM_FLAG) ;\
+	  cargo rustc --target-dir $(_PWD)/$$< -- --emit=llvm-bc -Clto
+	llc -march=riscv64 -target-abi=lp64d -mattr="+64bit,+m,+a,+f,+d,+c" -code-model=medium $$</riscv64gc-unknown-none-elf/debug/deps/*.bc -o $$@
+
+
+$$($(2)_crate_objs) : %.o : %.S
+	$(COMPILE) -c $$<
+$$($(2)_crate_libs) : %.a : %.o
+	$(AR) rcv $$@ $$(notdir $$<)
+	$(RANLIB) $$@
+
 $(2)_junk += $$($(2)_c_objs) $$($(2)_c_deps) $$($(2)_asm_objs)
 
 # Build a library for this subproject
 
-lib$(1).a : $$($(2)_c_objs) $$($(2)_asm_objs) librs.a
+lib$(1).a : $$($(2)_c_objs) $$($(2)_asm_objs) $$($(2)_crate_objs)
 	$(AR) rcv $$@ $$(notdir $$^)
 	$(RANLIB) $$@
