+++ sm/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "sm"
+name = "sm-rs"
 version = "0.1.0"
 edition = "2018"
 
 [lib]
+crate-type = ["staticlib"]
 path = "lib.rs"
 
 [dependencies]
