+++ sm/sm.mk.in
@@ -23,3 +23,5 @@ sm_asm_srcs =
 sm_test_srcs =
 
 sm_install_prog_srcs =
+
+sm_crates = sm-rs
